Se hicieron correciones en el código del formulario Tablas (Dashboard). Se agregaron Formularios (Pantallas): -Reporte -ReporteInfo -ReporteKevlar1 Estas pantallas se agregaron para la creación del reporte General Se hicieron cambios en el código del proyecto DataAccess y Domain

+++ AppMantenimiento/FrmMantenimientoCorrectivo.cs
@@ -164,7 +164,7 @@ namespace AppMantenimiento
 
         private void bttGuardarListado_Click(object sender, EventArgs e)
         {
-            listado.AgregarListado(correctivo.Id, (int)cmbListadoTrabajadores.SelectedValue, txtDescListado.Text, (int)nmrHorasTrabajadas.Value);
+            listado.AgregarListado(correctivo.Id, (int)cmbListadoTrabajadores.SelectedValue, txtDescListado.Text, (double)nmrHorasTrabajadas.Value);
             MessageBox.Show("Trabajador asignado correctamente.", "Mensaje", MessageBoxButtons.OK, MessageBoxIcon.Information);
             bttCerrarListado.Enabled = true;
         }
